requirements.txt
django-debug-toolbar
django
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.7
bs4
pprint
Faker
mixer